Consumable items support
Added support for consumable items. Now when the consumable item is used (by pressing E, selling items is still done by clicking Space), it either sells itself or spawns a new item in the Items list. The spawned new item is in Sellable type, so it can be sold afterward.